# api/swagger.yaml
openapi: 3.0.1
info:
  title: MZT Health Fitness App
  description: A webapp to assist a coach in managing their client's workout programs.
  version: 0.0.1
servers:
- url: /
paths:
...
  /program/assignProgramTemplate:
    post:
      description: Assigns program templates to program
      operationId: assignProgramTemplate
      requestBody:
        description: Assigns program templates to program
        required: true
        content:
          application/json:
            schema:
               $ref: '#/components/parameters/programTempIdParamAssign'
      responses:
        200:
          description: OK Program Info Obtained
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Program'
                x-content-type: application/json
        400:
          description: 'Error: Bad Request'
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: 'Error: Not Found'
      security:
      - basicAuth: []
      x-swagger-router-controller: program.controller
  /program/getProgramById/{program_id}:
    get:
      description: Gets all program information by program ID
      operationId: getProgramById
      parameters:
      - $ref: '#/components/parameters/programIdParamGet'
      responses:
        200:
          description: OK Program Info Obtained
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Program'
                x-content-type: application/json
        400:
          description: 'Error: Bad Request'
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: 'Error: Not Found'
      security:
      - basicAuth: []
      x-swagger-router-controller: program.controller
  /program/getProgramByCustomerId/{customer_id}:
    get:
      description: Gets all program information by customer ID
      operationId: getProgramByCustomerId
      parameters:
      - $ref: '#/components/parameters/customerIdParamGet'
      responses:
        200:
          description: OK Program Info Obtained
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Program'
                x-content-type: application/json
        400:
          description: 'Error: Bad Request'
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: 'Error: Not Found'
      security:
      - basicAuth: []
      x-swagger-router-controller: program.controller
  /program/getAllPrograms:
    get:
      description: Gets list of all programs
      operationId: getAllPrograms
      responses:
        200:
          description: OK Program Info Obtained
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Program'
                x-content-type: application/json
        400:
          description: 'Error: Bad Request'
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: 'Error: Not Found'
      security:
      - basicAuth: []
      x-swagger-router-controller: program.controller
  /program/updateProgram:
    post:
      description: Updates an existing program
      operationId: updateProgram
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Program'
      responses:
        200:
          description: OK Program Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Program'
                x-content-type: application/json
        400:
          description: 'Error: Bad Request'
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: 'Error: Not Found'
      security:
      - basicAuth: []
      x-swagger-router-controller: program.controller
  /program/deleteProgram:
    delete:
      description: Deletes an existing program
      operationId: deleteProgram
      parameters:
        - $ref: '#/components/parameters/programIdParamDel'
      responses:
        200:
          description: OK Program Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/parameters/programIdParamDel'
                x-content-type: application/json
        400:
          description: 'Error: Bad Request'
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: 'Error: Not Found'
      security:
      - basicAuth: []
      x-swagger-router-controller: program.controller
  /programTemplate/assignSessionTemplate:
    post:
      description: Assigns session templates to program template
      operationId: assignSessionTemplate
      requestBody:
        description: Assigns session templates to program template
        required: true
        content:
          application/json:
            schema:
               $ref: '#/components/parameters/sessionTempIdParamAssign'
      responses:
        200:
          description: OK Program Template Info Obtained
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgramTemplate'
                x-content-type: application/json
        400:
          description: 'Error: Bad Request'
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: 'Error: Not Found'
      security:
      - basicAuth: []
      x-swagger-router-controller: programTemplate.controller
  /programTemplate/getProgramTempById/{program_template_id}:
    get:
      description: Gets all program template information by program ID
      operationId: getProgramTempById
      parameters:
      - $ref: '#/components/parameters/programTempIdParamGet'
      responses:
        200:
          description: OK Program Template Info Obtained
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgramTemplate'
                x-content-type: application/json
        400:
          description: 'Error: Bad Request'
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: 'Error: Not Found'
      security:
      - basicAuth: []
      x-swagger-router-controller: programTemplate.controller
  /programTemplate/getAllProgramTemps:
    get:
      description: Gets list of all program templates
      operationId: getAllProgramTemps
      responses:
        200:
          description: OK Program Templates Info Obtained
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProgramTemplate'
                x-content-type: application/json
        400:
          description: 'Error: Bad Request'
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: 'Error: Not Found'
      security:
      - basicAuth: []
      x-swagger-router-controller: programTemplate.controller
  /programTemplate/addProgramTemp:
    post:
      description: Creates new program template
      operationId: addProgramTemp
      requestBody:
        description: Add a new program template
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProgramTemplate'
      responses:
        200:
          description: OK Program Template Info Obtained
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgramTemplate'
                x-content-type: application/json
        400:
          description: 'Error: Bad Request'
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: 'Error: Not Found'
      security:
      - basicAuth: []
      x-swagger-router-controller: programTemplate.controller
  /programTemplate/updateProgramTemp:
    post:
      description: Updates an existing program template
      operationId: updateProgramTemp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProgramTemplate'
      responses:
        200:
          description: OK Program Template Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgramTemplate'
                x-content-type: application/json
        400:
          description: 'Error: Bad Request'
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: 'Error: Not Found'
      security:
      - basicAuth: []
      x-swagger-router-controller: programTemplate.controller
  /programTemplate/deleteProgramTemp:
    delete:
      description: Deletes an existing program template
      operationId: deleteProgramTemp
      parameters:
        - $ref: '#/components/parameters/programTempIdParamDel'
      responses:
        200:
          description: OK Program Template Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/parameters/programTempIdParamDel'
                x-content-type: application/json
        400:
          description: 'Error: Bad Request'
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: 'Error: Not Found'
      security:
      - basicAuth: []
      x-swagger-router-controller: programTemplate.controller
...
  /session/getAllSessionsByProgId/{program_id}:
    get:
      description: Gets a session by program ID
      operationId: getAllSessionsByProgId
      parameters:
      - $ref: '#/components/parameters/programIdParamGet'
      responses:
        200:
          description: OK Session Info Obtained
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
                x-content-type: application/json
        400:
          description: 'Error: Bad Request'
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: 'Error: Not Found'
      security:
      - basicAuth: []
      x-swagger-router-controller: session.controller
  /session/getSessionById/{session_id}:
    get:
      description: Gets an session by ID
      operationId: getSessionById
      parameters:
      - $ref: '#/components/parameters/sessionIdParamGet'
      responses:
        200:
          description: OK Session Info Obtained
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
                x-content-type: application/json
        400:
          description: 'Error: Bad Request'
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: 'Error: Not Found'
      security:
      - basicAuth: []
      x-swagger-router-controller: session.controller
  /session/addSession:
    post:
      description: Creates new session
      operationId: addSession
      requestBody:
        description: Add a new session
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Session'
      responses:
        200:
          description: OK Session Info Obtained
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
                x-content-type: application/json
        400:
          description: 'Error: Bad Request'
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: 'Error: Not Found'
      security:
      - basicAuth: []
      x-swagger-router-controller: session.controller
  /session/updateSession:
    post:
      description: Updates an existing session
      operationId: updateSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Session'
      responses:
        200:
          description: OK Session Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
                x-content-type: application/json
        400:
          description: 'Error: Bad Request'
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: 'Error: Not Found'
      security:
      - basicAuth: []
      x-swagger-router-controller: session.controller
  /session/deleteSession:
    delete:
      description: Deletes an existing session
      operationId: deleteSession
      parameters:
        - $ref: '#/components/parameters/sessionIdParamDel'
      responses:
        200:
          description: OK session Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/parameters/sessionIdParamDel'
                x-content-type: application/json
        400:
          description: 'Error: Bad Request'
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: 'Error: Not Found'
      security:
      - basicAuth: []
      x-swagger-router-controller: session.controller
...
  /sessionTemplate/deleteSessionTemp:
    delete:
      description: Deletes an existing session template
      operationId: deleteSession
      parameters:
        - $ref: '#/components/parameters/sessionTempIdParamDel'
      responses:
        200:
          description: OK session template Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/parameters/sessionTempIdParamDel'
                x-content-type: application/json
        400:
          description: 'Error: Bad Request'
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: 'Error: Not Found'
      security:
      - basicAuth: []
      x-swagger-router-controller: sessionTemplate.controller
components:
  parameters:
    customerIdParamGet:
      name: customer_id
      in: path
      description: the id of the customer
      required: true
      style: simple
      explode: false
      schema:
        type: string
...
    programIdParamGet:
      name: program_id
      in: path
      description: the id of the program
      required: true
      style: simple
      explode: false
      schema:
        type: string
    programIdParamDel:
      name: program_id
      in: query
      description: the id of the program
      required: true
      style: simple
      explode: false
      schema:
        type: string
    programTempIdParamAssign:
      name: assignProgramTemp
      in: body
      description: the id of the program template
      required: true
      style: simple
      explode: false
      schema:
        type: Array
        items: {}
    programTempIdParamGet:
      name: program_template_id
      in: path
      description: the id of the program template
      required: true
      style: simple
      explode: false
      schema:
        type: string
    programTempIdParamDel:
      name: program_template_id
      in: query
      description: the id of the program template
      required: true
      style: simple
      explode: false
      schema:
        type: string
...
    sessionIdParamGet:
      name: session_id
      in: path
      description: the id of the session
      required: true
      style: simple
      explode: false
      schema:
        type: string
    sessionIdParamDel:
      name: session_id
      in: query
      description: the id of the session
      required: true
      style: simple
      explode: false
      schema:
        type: string
    sessionTempIdParamAssign:
      name: assignSessionTemp
      in: body
      description: the id of the session template
      required: true
      style: simple
      explode: false
      schema:
        type: Array
        items: {}
...
    sessionTempIdParamDel:
        name: session_template_id
        in: query
        description: the id of the session template
        required: true
        style: simple
        explode: false
        schema:
          type: string
  schemas:
...
    Exercise:
      type: object
      properties:
        exercise_id:
          type: object
        name:
          type: string
        description:
          type: string
        equipment_required:
          type: string
        exercise_type:
          type: string
        muscles_targeted:
          type: string
        set_break:
          type: integer
          description: the time in seconds to rest between each set
        repetitions:
          type: integer
          description: the count of one atomic set of movement in a set
        sets:
          type: integer
          description: a set is composed of a number of repititions
        exercise_est_duration:
          type: integer
          description: The time allocated to perform exercise
        session:
            type: Object
            description: The assigned session
        exercise_img_url:
            type: string
            description: image that demonstrates the exercise
        exercise_tag:
          type: array
          items: {}
      example:
        exercis_id: {}
        sets: 1
        exercise_tag:
        - ""
        - ""
        exercise_type: exerciseType
        name: name
        exercise_est_duration: 5
        description: description
        equipment_required: equipmentRequired
        set_break: 0
        repetitions: 6
        musclesTargeted: musclesTargeted
    Program:
      type: object
      properties:
        program_id:
          type: object
        title:
          type: string
        type:
          type: string
        description:
          type: string
        duration:
          type: string
          format: date
        customer_id:
          type: object
        sessions:
          type: array
          items: {}
        coach_id:
          type: object
        staus:
          type: string
    ProgramTemplate:
      type: object
      properties:
        program_template_id:
          type: object
        title:
          type: string
        type:
          type: string
        description:
          type: string
        duration:
          type: integer
        sessions:
          type: array
          items:
          - $ref: '#/components/schemas/SessionTemplate'
    Session:
        type: object
        properties:
            session_id:
                type: object
            session_name:
                type: string
            session_type:
                type: string
            session_duration:
                type: integer
            session_coach_notes:
                type: string
            session_customer_feedback:
                type: string
            exercises:
                type: array
                items:
                - $ref: '#/components/schemas/Exercise'
            session_tag:
                type: array
                items: {}
            measurement_date:
                type: string
                format: date
    SessionTemplate:
        type: object
        properties:
            session_template_id:
                type: object
            coach_id:
                type: object
            session_type:
                type: String
            session_coach_notes:
                type: String
            session_template_tag:
                type: array
                items: {}
            program_template_id:
                type: object