# service/program.service.js
const Program = require('../models/program.model.js');
const ProgramTemp = require('../models/programTemplate.model.js');
const SessionTemplate = require('../models/sessionTemplate.model.js');
require('../service/checkNull.js');

exports.assignProgramTemplate = function (coachId, custId, progTempId) {
    return new Promise(function (resolve, reject) {
        console.log('coach id: ' + coachId);
    console.log('customer id: ' + custId);
    console.log('program template id: ' + progTempId);
        ProgramTemp.findById(progTempId).then(progTemp => {
            if (!progTemp) {
                reject("Program template not found with id " + progTempId);
            }
            var Sessions = [];
            console.log(progTemp.sessions);
            for (var session in progTemp.sessions){
                console.log("session: "+session);
                console.log('progTemp.sessions[session]'+progTemp.sessions[session]);
            SessionTemplate.findById(progTemp.sessions[session]).then(sess => {
                    if (!sess) {
                        reject("Session templates not found with id " + progTemp.sessions[session]);
                    }
                console.log("[session]:" + session);
                Sessions.push(sess);
                console.log("Sessions[session]:" + Sessions[session]);
                console.log("Sessions: " + Sessions);
            });
        };
        console.log("Sessions: "+Sessions);
        const program = new Program({
            title: progTemp['title'],
            type: progTemp['type'],
            description: progTemp['description'],
            duration: progTemp['duration'],
            customer_id: custId,
            coach_id: coachId
        });
        // Save the program in the database
        program.save()
                .then(data => {
                    console.log(data);
        Program.findByIdAndUpdate(data._id, {'$push':{sessions: Sessions}}, {new : true})
                .then(program => {
                    if (!program) {
                        reject("Session not found with id " + data._id);
                    }
                    resolve(program);
                }).catch(err => {
            reject(err.message || "Some error occurred while creating the program.");
        });
    });
        }).catch(err => {
            if (err.kind === 'ObjectId') {
                reject("Session not found with id " + progTempId);
            }
            reject("Error retrieving session with id " + progTempId);
        });
    });
};

exports.findAll = function () {
    return new Promise(function (resolve, reject) {
        Program.find()
                .then(programs => {
                    resolve(programs);
                }).catch(err => {
            reject('IT WAS REJECTED');
        });
    });
};

exports.findOne = function (progId) {
    return new Promise(function (resolve, reject) {
        Program.findById(progId)
                .then(programs => {
                    resolve(programs);
                }).catch(err => {
            if (err.kind === 'ObjectId') {
                reject('Program Not Found By ID');
            }
            //TODO setup real error messages
            reject('INTERNAL ERROR');
        });
    });
};

exports.findProgramByCustomerId = function (custId) {
    return new Promise(function (resolve, reject) {
        var query = {customer_id: custId};
        Program.find(query)
                .then(programs => {
                    resolve(programs);
                }).catch(err => {
            if (err.kind === 'ObjectId') {
                reject('Program Not Found By Customer ID');
            }
            //TODO setup real error messages
            reject('INTERNAL ERROR');
        });
    });
};

// Update a program identified by the programId in the request
exports.update = function (prog) {
    return new Promise(function (resolve, reject) {
        // Validate Request
        if (!prog) {
            reject("Program content can not be empty");
        }
        var raw = {};
        raw = checkNull(raw, prog);
        // Find program and update its name with the request body
        Program.findByIdAndUpdate(prog['program_id'], {
            title: prog['title'] || "NA",
            discription: prog['discription'] || "NA"
        }, {new : true})
                .then(programs => {
                    if (!programs) {
                        reject("Program not found with id " + prog['program_id']);
                    }
                    resolve(programs);
                }).catch(err => {
            if (err.kind === 'ObjectId') {
                reject("Program not found with id " + prog['program_id']);
            }
            reject("Error updating program with id " + prog['program_id']);
        });
    });
};

// Delete a program with the specified programId in the request
exports.delete = function (program_id) {
    return new Promise(function (resolve, reject) {
        Program.findByIdAndRemove(program_id)
                .then(programs => {
                    if (!programs) {
                        reject("Program not found with id " + program_id);
                    }
                    resolve("Program deleted successfully!");
                }).catch(err => {
            if (err.kind === 'ObjectId' || err.name === 'NotFound') {
                reject("Program not found with id " + program_id);
            }
            reject("Could not delete program with id " + program_id);
        });
    });
};